+++ tgbotapi.behaviour_builder/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/behaviour_builder/expectations/WaitContentMessage.kt
@@ -13,17 +13,20 @@ import dev.inmo.tgbotapi.utils.RiskFeature
 import dev.inmo.tgbotapi.utils.lowLevelRiskFeatureMessage
 import kotlinx.coroutines.flow.Flow
 
+const val includeMediaGroupsDeprecationMessage = "includeMediaGroups is deprecated and its usage will not lead to any changes"
 typealias CommonMessageToCommonMessageMapper<T> = suspend CommonMessage<T>.() -> CommonMessage<T>?
 
 @RiskFeature(lowLevelRiskFeatureMessage)
 suspend inline fun <reified O : MessageContent> BehaviourContext.waitContentMessage(
     initRequest: Request<*>? = null,
-    includeMediaGroups: Boolean = true,
     noinline errorFactory: NullableRequestBuilder<*> = { null }
 ): Flow<CommonMessage<O>> = expectFlow(
     initRequest,
     errorFactory
 ) {
+    if (it !is BaseSentMessageUpdate) {
+        return@expectFlow emptyList()
+    }
     listOfNotNull((it.data as? CommonMessage<*>) ?.withContent<O>())
 }
 
@@ -44,114 +47,172 @@ internal inline fun <reified T : MessageContent> contentMessageConverter(
     if (content is T) this as CommonMessage<T> else null
 }
 
+@Deprecated(
+    includeMediaGroupsDeprecationMessage,
+    ReplaceWith("waitAnyContentMessage(initRequest, errorFactory)", "dev.inmo.tgbotapi.extensions.behaviour_builder.expectations.waitAnyContentMessage")
+)
 suspend fun BehaviourContext.waitContentMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = true
-) = waitContentMessage<MessageContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<MessageContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitAnyContentMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null },
+) = waitContentMessage<MessageContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitContactMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<ContactContent>(initRequest, false, errorFactory)
+) = waitContentMessage<ContactContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitDiceMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<DiceContent>(initRequest, false, errorFactory)
+) = waitContentMessage<DiceContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitGameMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<GameContent>(initRequest, false, errorFactory)
+) = waitContentMessage<GameContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitLocationMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<LocationContent>(initRequest, false, errorFactory)
+) = waitContentMessage<LocationContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitLiveLocationMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<LiveLocationContent>(initRequest, false, errorFactory)
+) = waitContentMessage<LiveLocationContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitStaticLocationMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<StaticLocationContent>(initRequest, false, errorFactory)
+) = waitContentMessage<StaticLocationContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitPollMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<PollContent>(initRequest, false, errorFactory)
+) = waitContentMessage<PollContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitTextMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<TextContent>(initRequest, false, errorFactory)
+) = waitContentMessage<TextContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitVenueMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<VenueContent>(initRequest, false, errorFactory)
+) = waitContentMessage<VenueContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitAudioMediaGroupContentMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = true
-) = waitContentMessage<AudioMediaGroupPartContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<AudioMediaGroupPartContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitAudioMediaGroupContentMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<AudioMediaGroupPartContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitDocumentMediaGroupContentMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = true
-) = waitContentMessage<DocumentMediaGroupPartContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<DocumentMediaGroupPartContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitDocumentMediaGroupContentMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<DocumentMediaGroupPartContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitMediaMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = false
-) = waitContentMessage<MediaContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<MediaContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitMediaMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<MediaContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitAnyMediaGroupContentMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = true
-) = waitContentMessage<MediaGroupPartContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<MediaGroupPartContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitAnyMediaGroupContentMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<MediaGroupPartContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitVisualMediaGroupContentMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = true
-) = waitContentMessage<VisualMediaGroupPartContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<VisualMediaGroupPartContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitVisualMediaGroupContentMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<VisualMediaGroupPartContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitTextedMediaContentMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = true
-) = waitContentMessage<TextedMediaContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<TextedMediaContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitTextedMediaContentMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<TextedMediaContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitAnimationMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<AnimationContent>(initRequest, false, errorFactory)
+) = waitContentMessage<AnimationContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitAudioMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = false
-) = waitContentMessage<AudioContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<AudioContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitAudioMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<AudioContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitDocumentMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = false
-) = waitContentMessage<DocumentContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<DocumentContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitDocumentMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<DocumentContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitPhotoMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = false
-) = waitContentMessage<PhotoContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<PhotoContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitPhotoMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<PhotoContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitStickerMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<StickerContent>(initRequest, false, errorFactory)
+) = waitContentMessage<StickerContent>(initRequest, errorFactory)
+@Deprecated(includeMediaGroupsDeprecationMessage)
 suspend fun BehaviourContext.waitVideoMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null },
-    includeMediaGroups: Boolean = false
-) = waitContentMessage<VideoContent>(initRequest, includeMediaGroups, errorFactory)
+    includeMediaGroups: Boolean
+) = waitContentMessage<VideoContent>(initRequest, errorFactory)
+suspend fun BehaviourContext.waitVideoMessage(
+    initRequest: Request<*>? = null,
+    errorFactory: NullableRequestBuilder<*> = { null }
+) = waitContentMessage<VideoContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitVideoNoteMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<VideoNoteContent>(initRequest, false, errorFactory)
+) = waitContentMessage<VideoNoteContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitVoiceMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<VoiceContent>(initRequest, false, errorFactory)
+) = waitContentMessage<VoiceContent>(initRequest, errorFactory)
 suspend fun BehaviourContext.waitInvoiceMessage(
     initRequest: Request<*>? = null,
     errorFactory: NullableRequestBuilder<*> = { null }
-) = waitContentMessage<InvoiceContent>(initRequest, false, errorFactory)
+) = waitContentMessage<InvoiceContent>(initRequest, errorFactory)