+++ tgbotapi.utils/src/commonMain/kotlin/dev/inmo/tgbotapi/extensions/utils/CommonMessageConversations.kt
@@ -1,6 +1,7 @@
 package dev.inmo.tgbotapi.extensions.utils
 
 import dev.inmo.tgbotapi.types.message.abstracts.*
+import dev.inmo.tgbotapi.types.message.content.MediaGroupContent
 import dev.inmo.tgbotapi.types.message.content.MessageContent
 import kotlinx.coroutines.flow.*
 
@@ -33,4 +34,11 @@ fun <MC : MessageContent, M : ContentMessage<MC>> Flow<M>.onlySentViaBot() = map
  */
 fun <MC : MessageContent, M : ContentMessage<MC>> Flow<M>.withoutSentViaBot() = filter {
     it !is PossiblySentViaBot || it.senderBot == null
+}
+
+/**
+ * Filter the messages and checking that incoming [ContentMessage.content] is not [MediaGroupContent]
+ */
+fun <MC : MessageContent, M : ContentMessage<MC>> Flow<M>.withoutMediaGroups() = filter {
+    it.content !is MediaGroupContent<*>
 }