+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 5.1.1
 
+* `Core`:
+    * Add opportunity to get user link with `makeUserLink`
+
 ## 5.1.0
 
 [Bot API 6.5](https://core.telegram.org/bots/api-changelog#february-3-2023) support